feat: memory, syscall constraints and recursion on RV64 (#493)
* checkpoint 1: syscall addr operation

* checkpoint 2: addr add operation refactor

* checkpoint 3: little bit of clippy

* checkpoint 4: memory operations

* checkpoint 5: syscall instr

* checkpoint 6: clippy

* checkpoint 7: memory global

* checkpoint 8: addr as 3 limbs in pv

* checkpoint 9: e2e recursion works

* checkpoint 10: clippy, trace area estimation

* chore: keccak controller

+++ crates/zkvm/entrypoint64/src/syscalls/halt.rs
@@ -32,7 +32,8 @@ pub extern "C" fn syscall_halt(exit_code: u8) -> ! {
         // will be used to verify that the provided public values digest matches the one
         // computed by the program.
         for i in 0..PV_DIGEST_NUM_WORDS {
-            let word = u32::from_le_bytes(pv_digest_bytes[i * 4..(i + 1) * 4].try_into().unwrap());
+            let word =
+                u32::from_le_bytes(pv_digest_bytes[i * 4..(i + 1) * 4].try_into().unwrap()) as u64;
             asm!("ecall", in("t0") crate::syscalls::COMMIT, in("a0") i, in("a1") word);
         }
